Sincronización de la biblioteca de Steam
Se arregló la funcionalidad para vincular los juegos de Steam (se usó un array provisionalmente), por lo que puede que no funcione correctamente.

 - Se modificó la base de datos, concretamente igdb_id de la tabla games para aceptar un bigint en lugar de un int

 - Se modificó la página Synchronize.jsx para que no realice la sincronización de la cuenta si ya está sincronizada. Para ello se creó el endpoint checkSteamStatus.php que comprueba si ya existe un provider de steam para ese usuario

 -  Se modificó la función searchGameByName() en gameFunctions.php para que devolviera sólo un resultado

 - Se añadieron errorlogs en la función getSteamGamesNames() en steamFunctions.php, que ayudaron a resolver problemas de conexión con la API de Steam (se pueden eliminar en un futuro)

 - Se modificó la consulta de la función  searchIgdbGameByName() en igdbFunctions.php para que recogiera la url del cover correctamente (lo que ayudó enormemente en la sincronización)

 - Por último se rehizo el endpoint getSteamLib.php para que funcionase y devolviese feedback correcto al frontend

+++ backend/database/gamelib.sql
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS games (
     -- Internal ID
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     -- IGDB API ID
-    igdb_id INT NOT NULL,
+    igdb_id BIGINT NOT NULL,
     -- JSON to creates the Image URL
     cover VARCHAR(255) NOT NULL,
     -- Title
